feat: add worker module
The worker module is responsible for creating an Auto Scaling Group of
Boundary workers.

The `worker` module is dependent on the `controller` module. In order to
allow the workers to connect to the controller hosts, the `worker`
module requires the `ip_addresses` variable.

The value of the `ip_addresses` variable is a list of strings, where
each value is the private IP address of each controller EC2 instance
instance in the Auto Scaling Group.

When the `controller` module initializes the PostgreSQL database, it
writes the credentials for the `admin` user to stdout.

The `boundary` module will grep the contents of the
`/var/log/cloud-init-output.log` file and if it contains the text
"Initial auth information", then it writes the contents to an S3 bucket.

Writing the contents to an S3 bucket ensures that the admin credentials
are always available to the user, even if the Auto Scaling Group is
re-created.