chore: remove SSH ingress rules
The worker and controller security groups included a rule to allow
ingress SSH traffic from anywhere.

I am temporarily removing the permissive rules for security purposes.

Eventually I plan to implement functionality to create a bastion host
that will serve as the entry point to the worker and controller
instances.